requirements.txt
Django==1.5.2
South==0.8.1
coverage==3.6
dj-database-url==0.2.1
gunicorn==0.17.4
psycopg2==2.5
wsgiref==0.1.2
dj-static==0.0.5
django-reversion==1.8.0
django-axes==1.3.4
ffs==0.0.7.6
letter==0.4.1
supervisor
-e git://github.com/srosro/django-test-coverage#egg=django-test-coverage
django-debug-toolbar
-e git://github.com/openhealthcare/opal.git@renal#egg=opal
